requirements.txt
streamlit>=1.30.0
Pillow>=10.0.0
rembg>=2.0.50
openai>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
dashscope>=1.23.4
